+++ src/lib/Tabs/assets/Icons.tsx
